fix(ai): relocate AgentId registry to lib/db
Once an action imported AGENT_IDS from @/lib/ai, vitest.setup.ts's registerAllDomains pulled the whole @/lib/ai barrel — the 'ai' SDK plus transport/fetch and model — into every unit test's setup graph, loading those modules before the test files' vi.mocks registered and silently breaking them. AGENT_IDS is config data that indexes app_settings.assignments, so lib/db is its proper home; lib/ai re-exports it so its public surface is unchanged.
